% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xml_list.emeSchemeSet.R
\name{as_xml_list.emeSchemeSet}
\alias{as_xml_list.emeSchemeSet}
\title{Convert \code{emeSchemeSet} to an \code{xml_list} object}
\usage{
\method{as_xml_list}{emeSchemeSet}(x, output = "metadata", ...)
}
\arguments{
\item{x}{object to be converted.}

\item{output}{specifies the content and format of the exported xml.
\describe{
  \item{"metadata" : }{export of the metadata only with no format attributes}
  \item{"complete" : }{export tof the complete sheme, i.e. "metadata" plus the scheme definition. This is a self contained format which contains all attributes.}
}}

\item{...}{additional arguments for methods}
}
\description{
Convert \code{emeSchemeSet} to an \code{xml_list} object
}
