\name{emeScheme-package}
\alias{emeScheme-package}
\alias{emeScheme}
\docType{package}
\title{
\packageTitle{emeScheme}
}
\description{
\packageDescription{emeScheme}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{emeScheme}
\packageIndices{emeScheme}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{emeScheme}

Maintainer: \packageMaintainer{emeScheme}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
