% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emeScheme_to_xml.R
\name{emeScheme_to_xml}
\alias{emeScheme_to_xml}
\title{Convert file name or \code{emeScheme} object to one xml per data file}
\usage{
emeScheme_to_xml(x, file, output = "metadata")
}
\arguments{
\item{x}{\code{emeScheme} object to be converted.}

\item{file}{empty, \code{NULL} or file name. See details below}

\item{output}{either \code{"metadata"} for export of metadata only or
\code{"complete"} for export including classes et al.}
}
\value{
dependent on the value of \code{file}. See Details
}
\description{
Converts an \code{emeScheme} object into xml and optionally saves it to a file.
}
\details{
Depending on the value of \code{file}, the function returns different results:
\itemize{
  \item{\code{file} not given: }{an \code{XMLNode} object}
  \item{\code{file = NULL}: }{a string characterisation of the \code{XMLNode} object}
  \item{\code{file = fileName}: }{saves the string characterisation of the \code{XMLNode} object to an xml file}
}
The actual implementation is inspired by David LeBauer, Carl Davidson, Rob Kooper. See \url{https://stackoverflow.com/a/27865512/632423}
}
\examples{
emeScheme_to_xml(emeScheme_example)
## x is a list containing all the emeSchemes for each data file

<<<<<<< HEAD
emeScheme_split(emeScheme_example, saveAsType = "xml", path = tempdir())
=======
emeScheme_to_xml(emeScheme_example, file = tempdir())
>>>>>>> 96eb085c5de3488ded931b4fa6e2f751be35694a
## saves the resulting object as xml into the tmpdir()
}
