% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xml_list.emeScheme.R
\name{as_xml_list.emeSchemeSet}
\alias{as_xml_list.emeSchemeSet}
\title{Convert \code{emeSchemeSet} to an \code{xml_list} object}
\usage{
\method{as_xml_list}{emeSchemeSet}(x, output = "metadata", ...)
}
\description{
Convert \code{emeSchemeSet} to an \code{xml_list} object
}
